MyCart: component refactorered/separated MyCart Main to MyCartSingleItem
